% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_shots.R
\name{extract_shots}
\alias{extract_shots}
\title{Extract shot data for Spanish FEB basketball leagues}
\usage{
extract_shots(game_ids)
}
\arguments{
\item{game_ids}{Game IDs from Baloncesto en Vivo}
}
\value{
A data frame with x, y coordinates
}
\description{
Extract shot data for Spanish FEB basketball leagues
}
\examples{
ids <- 2010208:2010216
week17_2018 <- extract_shots(ids)
}
