% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shotchart.R
\name{plot_shotchart}
\alias{plot_shotchart}
\title{Plot shot charts}
\usage{
plot_shotchart(df, ...)
}
\arguments{
\item{df}{data frame with at least two coordinate columns named x and y}

\item{...}{arguments specified to geom_point(aes(x, y, ...))}
}
\value{
A ggplot object
}
\description{
Plot shot charts
}
\examples{
library(ggplot2)
ids <- 2010208:2010216
week17_2018 <- extract_shots(ids)
plot_shotchart(week17_2018)
plot_shotchart(week17_2018, color = made) + facet_wrap(~game)
}
