% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_raw_list.R
\name{scrape_raw_list}
\alias{scrape_raw_list}
\title{Scrape raw list object with all the shooting data from Baloncesto en Vivo}
\usage{
scrape_raw_list(game_ids)
}
\arguments{
\item{game_ids}{Game IDs}
}
\value{
A list object
}
\description{
Scrape raw list object with all the shooting data from Baloncesto en Vivo
}
\examples{
bad_ids <- c(2010215, 2010150, 2010216, 2010147)
good_ids <- 2010213:2010216
good_requests <- scrape_raw_list(good_ids)
bad_request <- scrape_raw_list(bad_ids)
}
