% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shots}
\alias{shots}
\title{Shooting data from LEB Oro}
\format{A data frame with variables:
\describe{
  \item{x, y}{x, y coordinates of the shot}
  \item{time}{Time elapsed since begining of game}
  ...
}}
\source{
\url{http://baloncestoenvivo.feb.es/}
}
\usage{
shots
}
\description{
Scraped data from Baloncesto en Vivo from week 17 of LEB Oro 2018/2019
}
\keyword{datasets}
